feat: reorganize schoolIcon components
