+++ src/common/components/CustomIcon/SchoolIcon/BaseIcon.tsx
@@ -1,7 +1,7 @@
-import { CustomIcon, type CustomIconProps } from "./CustomIcon";
-import { NTUIcon } from "@/common/components/CustomIcon/NTUIcon";
-import { NUSIcon } from "@/common/components/CustomIcon/NUSIcon";
-import { SMUIcon } from "@/common/components/CustomIcon/SMUIcon";
+import { CustomIcon, type CustomIconProps } from "../CustomIcon";
+import { NTUIcon } from "@/common/components/CustomIcon/SchoolIcon/NTUIcon";
+import { NUSIcon } from "@/common/components/CustomIcon/SchoolIcon/NUSIcon";
+import { SMUIcon } from "@/common/components/CustomIcon/SchoolIcon/SMUIcon";
 
 export interface SchoolIconProps extends CustomIconProps {
   school: "NTU" | "SMU" | "NUS";
