+++ src/common/components/CustomIcon/SchoolIcon/NTUIcon.tsx
