+++ src/common/components/CustomIcon/SchoolIcon/NUSIcon.tsx
