+++ src/common/components/CustomIcon/SchoolIcon/SMUIcon.tsx
