+++ src/common/components/CustomIcon/SchoolIcon/index.ts
@@ -1,0 +1,1 @@
+export * from "./BaseIcon";
