+++ src/common/components/CustomIcon/index.ts
@@ -9,4 +9,4 @@ export * from "./MinusIcon";
 export * from "./StarLineAltIcon";
 export * from "./WarningCircleIcon";
 export * from "./XCloseIcon";
-export * from "./SchoolIcon";
+export * from "./SchoolIcon/BaseIcon";
